+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
  - Hide second navbar by default
  - Use javascript promises for all HTTP requests ([#53](https://github.com/pwyf/iati-org-viz/issues/53))
  - Improve error handling throughout, including handling invalid org files ([#48](https://github.com/pwyf/iati-org-viz/issues/48))
+ - Use IATI codelists as a reference for codelist codes used ([#38](https://github.com/pwyf/iati-org-viz/issues/38))
 
 ### Fixed